requirements.txt
appdirs==1.4.3
argcomplete==1.11.1
importlib-metadata==1.5.0
paho-mqtt==1.5.0
Pillow==7.0.0
pkg-resources==0.0.0
pyserial==3.4
python-escpos==2.2.0
pyusb==1.0.2
PyYAML==5.3
qrcode==6.1
six==1.14.0
zipp==3.0.0
